database/databaserequirements.txt
Django
djangorestframework
Pillow
django-cleanup